@@ -9,27 +9,16 @@ use Spatie\Backup\Tasks\Backup\BackupJob;
 
 class Backup
 {
-    /** @var \Illuminate\Contracts\Filesystem\Filesystem */
-    protected $disk;
+    protected bool $exists = true;
 
-    /** @var string */
-    protected $path;
+    protected ?Carbon $date = null;
 
-    /** @var bool */
-    protected $exists;
+    protected ?int $size = null;
 
-    /** @var Carbon */
-    protected $date;
-
-    /** @var int */
-    protected $size;
-
-    public function __construct(Filesystem $disk, string $path)
-    {
-        $this->disk = $disk;
-        $this->path = $path;
-        $this->exists = true;
-    }
+    public function __construct(
+        protected Filesystem $disk,
+        protected string $path,
+    ) {}
 
     public function disk(): Filesystem
     {
@@ -54,11 +43,10 @@ class Backup
     {
         if ($this->date === null) {
             try {
-                // try to parse the date from the filename
                 $basename = basename($this->path);
+
                 $this->date = Carbon::createFromFormat(BackupJob::FILENAME_FORMAT, $basename);
-            } catch (InvalidArgumentException $e) {
-                // if that fails, ask the (remote) filesystem
+            } catch (InvalidArgumentException) {
                 $this->date = Carbon::createFromTimestamp($this->disk->lastModified($this->path));
             }
         }
@@ -66,10 +54,7 @@ class Backup
         return $this->date;
     }
 
-    /**
-     * Get the size in bytes.
-     */
-    public function size(): float
+    public function sizeInBytes(): float
     {
         if ($this->size === null) {
             if (! $this->exists()) {
@@ -87,14 +72,14 @@ class Backup
         return $this->disk->readStream($this->path);
     }
 
-    public function delete()
+    public function delete(): void
     {
         if (! $this->disk->delete($this->path)) {
             consoleOutput()->error("Failed to delete backup `{$this->path}`.");
-            
+
             return;
         }
-        
+
         $this->exists = false;
 
         consoleOutput()->info("Deleted backup `{$this->path}`.");
