@@ -10,20 +10,15 @@ use Spatie\Backup\Exceptions\InvalidBackupDestination;
 
 class BackupDestination
 {
-    /** @var \Illuminate\Contracts\Filesystem\Filesystem */
-    protected $disk;
+    protected ?Filesystem $disk;
 
-    /** @var string */
-    protected $diskName;
+    protected string $diskName;
 
-    /** @var string */
-    protected $backupName;
+    protected string $backupName;
 
-    /** @var Exception */
-    public $connectionError;
+    public ?Exception $connectionError = null;
 
-    /** @var null|\Spatie\Backup\BackupDestination\BackupCollection */
-    protected $backupCollectionCache = null;
+    protected ?BackupCollection $backupCollectionCache = null;
 
     public function __construct(Filesystem $disk = null, string $backupName, string $diskName)
     {
@@ -31,7 +26,7 @@ class BackupDestination
 
         $this->diskName = $diskName;
 
-        $this->backupName = preg_replace('/[^a-zA-Z0-9.]/', '-', $backupName);
+        $this->backupName = (string)preg_replace('/[^a-zA-Z0-9.]/', '-', $backupName);
     }
 
     public function disk(): Filesystem
@@ -50,7 +45,7 @@ class BackupDestination
             return 'unknown';
         }
 
-        $adapterClass = get_class($this->disk->getDriver()->getAdapter());
+        $adapterClass = $this->disk->getDriver()->getAdapter()::class;
 
         $filesystemType = last(explode('\\', $adapterClass));
 
@@ -72,7 +67,7 @@ class BackupDestination
         }
     }
 
-    public function write(string $file)
+    public function write(string $file): void
     {
         if (! is_null($this->connectionError)) {
             throw InvalidBackupDestination::connectionError($this->diskName);
