@@ -9,8 +9,6 @@ class BackupDestinationFactory
     public static function createFromArray(array $config): Collection
     {
         return collect($config['destination']['disks'])
-            ->map(function ($filesystemName) use ($config) {
-                return BackupDestination::create($filesystemName, $config['name']);
-            });
+            ->map(fn ($filesystemName) => BackupDestination::create($filesystemName, $config['name']));
     }
 }
