@@ -6,11 +6,7 @@ use Spatie\Backup\Tasks\Backup\Manifest;
 
 class BackupManifestWasCreated
 {
-    /** @var \Spatie\Backup\Tasks\Backup\Manifest */
-    public $manifest;
-
-    public function __construct(Manifest $manifest)
-    {
-        $this->manifest = $manifest;
-    }
+    public function __construct(
+        public Manifest $manifest,
+    ) {}
 }
