@@ -6,11 +6,7 @@ use Spatie\Backup\BackupDestination\BackupDestination;
 
 class BackupWasSuccessful
 {
-    /** @var \Spatie\Backup\BackupDestination\BackupDestination */
-    public $backupDestination;
-
-    public function __construct(BackupDestination $backupDestination)
-    {
-        $this->backupDestination = $backupDestination;
-    }
+    public function __construct(
+        public BackupDestination $backupDestination,
+    ) {}
 }
