@@ -4,11 +4,7 @@ namespace Spatie\Backup\Events;
 
 class BackupZipWasCreated
 {
-    /** @var string */
-    public $pathToZip;
-
-    public function __construct(string $pathToZip)
-    {
-        $this->pathToZip = $pathToZip;
-    }
+    public function __construct(
+        public string $pathToZip,
+    ) {}
 }
