@@ -6,11 +6,7 @@ use Spatie\Backup\Tasks\Monitor\BackupDestinationStatus;
 
 class UnhealthyBackupWasFound
 {
-    /** @var \Spatie\Backup\Tasks\Monitor\BackupDestinationStatus */
-    public $backupDestinationStatus;
-
-    public function __construct(BackupDestinationStatus $backupDestinationStatus)
-    {
-        $this->backupDestinationStatus = $backupDestinationStatus;
-    }
+    public function __construct(
+        public BackupDestinationStatus $backupDestinationStatus
+    ) {}
 }
