@@ -26,8 +26,8 @@ class BackupJobFactory
 
     protected static function createDbDumpers(array $dbConnectionNames): Collection
     {
-        return collect($dbConnectionNames)->mapWithKeys(function (string $dbConnectionName) {
-            return [$dbConnectionName => DbDumperFactory::createFromConnection($dbConnectionName)];
-        });
+        return collect($dbConnectionNames)->mapWithKeys(
+            fn (string $dbConnectionName) => [$dbConnectionName => DbDumperFactory::createFromConnection($dbConnectionName)]
+        );
     }
 }
