@@ -3,12 +3,12 @@
 namespace Spatie\Backup\Tasks\Backup;
 
 use Countable;
+use Generator;
 use SplFileObject;
 
 class Manifest implements Countable
 {
-    /** @var string */
-    protected $manifestPath;
+    protected string $manifestPath;
 
     public static function create(string $manifestPath): self
     {
@@ -27,12 +27,7 @@ class Manifest implements Countable
         return $this->manifestPath;
     }
 
-    /**
-     * @param array|string $filePaths
-     *
-     * @return $this
-     */
-    public function addFiles($filePaths): self
+    public function addFiles(array|string|Generator $filePaths): self
     {
         if (is_string($filePaths)) {
             $filePaths = [$filePaths];
@@ -47,10 +42,7 @@ class Manifest implements Countable
         return $this;
     }
 
-    /**
-     * @return \Generator|string[]
-     */
-    public function files()
+    public function files(): Generator|array
     {
         $file = new SplFileObject($this->path());
 
