@@ -8,34 +8,21 @@ use Spatie\Backup\BackupDestination\BackupDestination;
 
 abstract class CleanupStrategy
 {
-    /** @var \Illuminate\Contracts\Config\Repository */
-    protected $config;
+    protected BackupDestination $backupDestination;
 
-    /** @var \Spatie\Backup\BackupDestination\BackupDestination */
-    protected $backupDestination;
-
-    public function __construct(Repository $config)
-    {
-        $this->config = $config;
-    }
+    public function __construct(
+        protected Repository $config,
+    ) {}
 
     abstract public function deleteOldBackups(BackupCollection $backups);
 
-    /**
-     * @param \Spatie\Backup\BackupDestination\BackupDestination $backupDestination
-     *
-     * @return $this
-     */
-    public function setBackupDestination(BackupDestination $backupDestination)
+    public function setBackupDestination(BackupDestination $backupDestination): self
     {
         $this->backupDestination = $backupDestination;
 
         return $this;
     }
 
-    /**
-     * @return \Spatie\Backup\BackupDestination\BackupDestination
-     */
     public function backupDestination(): BackupDestination
     {
         return $this->backupDestination;
