@@ -6,18 +6,10 @@ use Carbon\Carbon;
 
 class Period
 {
-    /** @var \Carbon\Carbon */
-    protected $startDate;
-
-    /** @var \Carbon\Carbon */
-    protected $endDate;
-
-    public function __construct(Carbon $startDate, Carbon $endDate)
-    {
-        $this->startDate = $startDate;
-
-        $this->endDate = $endDate;
-    }
+    public function __construct(
+        protected Carbon $startDate,
+        protected Carbon $endDate
+    ) {}
 
     public function startDate(): Carbon
     {
