@@ -11,8 +11,7 @@ use Spatie\Backup\Tasks\Cleanup\Period;
 
 class DefaultStrategy extends CleanupStrategy
 {
-    /** @var \Spatie\Backup\BackupDestination\Backup */
-    protected $newestBackup;
+    protected ?Backup $newestBackup = null;
 
     public function deleteOldBackups(BackupCollection $backups)
     {
@@ -22,9 +21,8 @@ class DefaultStrategy extends CleanupStrategy
         $dateRanges = $this->calculateDateRanges();
 
         $backupsPerPeriod = $dateRanges->map(function (Period $period) use ($backups) {
-            return $backups->filter(function (Backup $backup) use ($period) {
-                return $backup->date()->between($period->startDate(), $period->endDate());
-            });
+            return $backups
+                ->filter(fn (Backup $backup) => $backup->date()->between($period->startDate(), $period->endDate()));
         });
 
         $backupsPerPeriod['daily'] = $this->groupByDateFormat($backupsPerPeriod['daily'], 'Ymd');
@@ -73,9 +71,7 @@ class DefaultStrategy extends CleanupStrategy
 
     protected function groupByDateFormat(Collection $backups, string $dateFormat): Collection
     {
-        return $backups->groupBy(function (Backup $backup) use ($dateFormat) {
-            return $backup->date()->format($dateFormat);
-        });
+        return $backups->groupBy(fn (Backup $backup) => $backup->date()->format($dateFormat));
     }
 
     protected function removeBackupsForAllPeriodsExceptOne(Collection $backupsPerPeriod)
@@ -84,16 +80,16 @@ class DefaultStrategy extends CleanupStrategy
             $groupedBackupsByDateProperty->each(function (Collection $group) {
                 $group->shift();
 
-                $group->each->delete();
+                $group->each(fn (Backup $backup) => $backup->delete());
             });
         });
     }
 
     protected function removeBackupsOlderThan(Carbon $endDate, BackupCollection $backups)
     {
-        $backups->filter(function (Backup $backup) use ($endDate) {
-            return $backup->exists() && $backup->date()->lt($endDate);
-        })->each->delete();
+        $backups
+            ->filter(fn (Backup $backup) => $backup->exists() && $backup->date()->lt($endDate))
+            ->each(fn (Backup $backup) => $backup->delete());
     }
 
     protected function removeOldBackupsUntilUsingLessThanMaximumStorage(BackupCollection $backups)
@@ -105,12 +101,12 @@ class DefaultStrategy extends CleanupStrategy
         $maximumSize = $this->config->get('backup.cleanup.default_strategy.delete_oldest_backups_when_using_more_megabytes_than')
             * 1024 * 1024;
 
-        if (($backups->size() + $this->newestBackup->size()) <= $maximumSize) {
+        if (($backups->size() + $this->newestBackup->sizeInBytes()) <= $maximumSize) {
             return;
         }
         $oldest->delete();
 
-        $backups = $backups->filter->exists();
+        $backups = $backups->filter(fn (Backup $backup) => $backup->exists());
 
         $this->removeOldBackupsUntilUsingLessThanMaximumStorage($backups);
     }
