@@ -7,18 +7,10 @@ use Spatie\Backup\Exceptions\InvalidHealthCheck;
 
 class HealthCheckFailure
 {
-    /** @var \Spatie\Backup\Tasks\Monitor */
-    protected $healthCheck;
-
-    /** @var \Exception */
-    protected $exception;
-
-    public function __construct(HealthCheck $healthCheck, Exception $exception)
-    {
-        $this->healthCheck = $healthCheck;
-
-        $this->exception = $exception;
-    }
+    public function __construct(
+        protected HealthCheck $healthCheck,
+        protected Exception $exception
+    ) {}
 
     public function healthCheck(): HealthCheck
     {
