@@ -8,15 +8,11 @@ use Spatie\Backup\Tasks\Monitor\HealthCheck;
 
 class MaximumAgeInDays extends HealthCheck
 {
-    /** @var int */
-    protected $days;
+    public function __construct(
+        protected int $days = 1
+    ) {}
 
-    public function __construct($days = 1)
-    {
-        $this->days = $days;
-    }
-
-    public function checkHealth(BackupDestination $backupDestination)
+    public function checkHealth(BackupDestination $backupDestination): void
     {
         $this->failIf(
             $this->hasNoBackups($backupDestination),
@@ -31,12 +27,12 @@ class MaximumAgeInDays extends HealthCheck
         );
     }
 
-    protected function hasNoBackups(BackupDestination $backupDestination)
+    protected function hasNoBackups(BackupDestination $backupDestination): bool
     {
         return $backupDestination->backups()->isEmpty();
     }
 
-    protected function isTooOld(Backup $backup)
+    protected function isTooOld(Backup $backup): bool
     {
         if (is_null($this->days)) {
             return false;
