@@ -8,15 +8,11 @@ use Spatie\Backup\Tasks\Monitor\HealthCheck;
 
 class MaximumStorageInMegabytes extends HealthCheck
 {
-    /** @var int */
-    protected $maximumSizeInMegaBytes;
+    public function __construct(
+        protected int $maximumSizeInMegaBytes = 5000
+    )  {}
 
-    public function __construct(int $maximumSizeInMegaBytes = 5000)
-    {
-        $this->maximumSizeInMegaBytes = $maximumSizeInMegaBytes;
-    }
-
-    public function checkHealth(BackupDestination $backupDestination)
+    public function checkHealth(BackupDestination $backupDestination): void
     {
         $usageInBytes = $backupDestination->usedStorage();
 
