@@ -7,11 +7,9 @@ use Spatie\Backup\Tasks\Backup\Manifest;
 
 class ManifestTest extends TestCase
 {
-    /** @var string */
-    protected $pathToManifest;
+    protected string $pathToManifest;
 
-    /** @var \Spatie\Backup\Tasks\Backup\Manifest */
-    protected $manifest;
+    protected Manifest $manifest;
 
     public function setUp(): void
     {
@@ -101,8 +99,6 @@ class ManifestTest extends TestCase
 
     protected function getTestFiles(): array
     {
-        return collect(range(1, 3))->map(function (int $number) {
-            return $this->getStubDirectory()."/file{$number}.txt";
-        })->toArray();
+        return collect(range(1, 3))->map(fn (int $number) => $this->getStubDirectory()."/file{$number}.txt")->toArray();
     }
 }
