@@ -9,10 +9,8 @@ use Spatie\Backup\Tasks\Backup\BackupJobFactory;
 
 class BackupCommand extends BaseCommand
 {
-    /** @var string */
     protected $signature = 'backup:run {--filename=} {--only-db} {--db-name=*} {--only-files} {--only-to-disk=} {--disable-notifications} {--timeout=}';
 
-    /** @var string */
     protected $description = 'Run the backup.';
 
     public function handle()
@@ -69,8 +67,14 @@ class BackupCommand extends BaseCommand
 
     protected function guardAgainstInvalidOptions()
     {
-        if ($this->option('only-db') && $this->option('only-files')) {
-            throw InvalidCommand::create('Cannot use `only-db` and `only-files` together');
+        if (! $this->option('only-db')) {
+            return;
         }
+
+        if (! $this->option('only-files')) {
+            return;
+        }
+
+        throw InvalidCommand::create('Cannot use `only-db` and `only-files` together');
     }
 }
