@@ -7,16 +7,8 @@ use Spatie\Backup\BackupDestination\BackupDestination;
 
 class CleanupHasFailed
 {
-    /** @var \Exception */
-    public $exception;
-
-    /** @var \Spatie\Backup\BackupDestination\BackupDestination|null */
-    public $backupDestination;
-
-    public function __construct(Exception $exception, BackupDestination $backupDestination = null)
-    {
-        $this->exception = $exception;
-
-        $this->backupDestination = $backupDestination;
-    }
+    public function __construct(
+        Exception $exception,
+        BackupDestination $backupDestination = null,
+    ) {}
 }
