@@ -7,8 +7,7 @@ use Illuminate\Contracts\Filesystem\Filesystem;
 
 class File
 {
-    /** @var array */
-    protected static $allowedMimeTypes = [
+    protected static array $allowedMimeTypes = [
         'application/zip',
         'application/x-zip',
         'application/x-gzip',
@@ -28,18 +27,18 @@ class File
         return pathinfo($path, PATHINFO_EXTENSION) === 'zip';
     }
 
-    protected function hasAllowedMimeType(?Filesystem $disk, string $path)
+    protected function hasAllowedMimeType(?Filesystem $disk, string $path): bool
     {
         return in_array($this->mimeType($disk, $path), self::$allowedMimeTypes);
     }
 
-    protected function mimeType(?Filesystem $disk, string $path)
+    protected function mimeType(?Filesystem $disk, string $path): bool|string
     {
         try {
             if ($disk && method_exists($disk, 'mimeType')) {
                 return $disk->mimeType($path) ?: false;
             }
-        } catch (Exception $exception) {
+        } catch (Exception) {
             // Some drivers throw exceptions when checking mime types, we'll
             // just fallback to `false`.
         }
