@@ -8,17 +8,17 @@ class Notifiable
 {
     use NotifiableTrait;
 
-    public function routeNotificationForMail()
+    public function routeNotificationForMail(): string
     {
         return config('backup.notifications.mail.to');
     }
 
-    public function routeNotificationForSlack()
+    public function routeNotificationForSlack(): string
     {
         return config('backup.notifications.slack.webhook_url');
     }
 
-    public function getKey()
+    public function getKey(): int
     {
         return 1;
     }
